backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.0.0
drf-extra-fields==3.4.0
django-filter==23.1
regex==2024.7.24
short_url==1.2.2
pyshorteners==1.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.3 
environs==11.0.0
flake8==6.0.0 
flake8-isort==6.0.0